Added user authentication system

+++ app/Controller/UsersController.php
@@ -10,19 +10,30 @@
 
 
 		public function login() {
+
 	        $this->layout = 'login';
 
+	        if ($this->Auth->loggedIn()) {
+	        	return $this->redirect($this->Auth->redirect()); 
+	        }
+
 	        if( $this->request->isPost() ) {
+
 	            if ( $this->Auth->login() ) {
 					$this->Cookie->destroy();
-					//$this->loginInfo($this->Session->read('Auth.User.id'));
 	                $this->redirect($this->Auth->redirect()); 
 	            } else {
 	                $this->Session->setFlash('Invalid username or password', 'default', array( 'class' => 'alert alert-danger' ) );
 	            }
+
 	        }
 
 	        $this->remove_all_cookies();
 	        $this->set('title_for_layout', 'OSAS - LogIn' );  
+	    }
+
+
+	    public function logout() {
+	    	return $this->redirect($this->Auth->logout());
 	    }
 	}
